+++ CSRequest.Test/ResponseExtensions_Test.cs
@@ -1,7 +1,5 @@
 ﻿using CSRequest.Test.Helpers;
 using FluentAssertions;
-using System;
-using System.Net.Http;
 using System.Threading.Tasks;
 using Xunit;
