+++ CSRequest/Internal/HeaderRequestTransform.cs
@@ -1,8 +1,4 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using System.Net.Http;
-using System.Threading.Tasks;
-using System.Reflection;
+﻿using System.Net.Http;
 
 namespace CSRequest.Internal
 {
