+++ CSRequest/Internal/UrlSegmentsRequestTransform.cs
@@ -1,5 +1,4 @@
-﻿using Newtonsoft.Json;
-using System;
+﻿using System;
 using System.Net.Http;
 using System.Text;
 
@@ -24,22 +23,6 @@ namespace CSRequest.Internal
             }
             sb.AppendJoin('/', segments);
             msg.RequestUri = new Uri(sb.ToString());
-        }
-    }
-
-    public class JsonContentRequestTransform : IRequestTransform
-    {
-        private readonly object data;
-
-        public JsonContentRequestTransform(object data)
-        {
-            this.data = data;
-        }
-
-        public void Transform(HttpRequestMessage msg)
-        {
-            var json = JsonConvert.SerializeObject(data);
-            msg.Content = new StringContent(json, Encoding.UTF8, "application/json");
         }
     }
 }
